add support for react-native firestore
