+++ index.d.ts
@@ -2,3 +2,4 @@ export { FakeFirestore } from './mocks/firestore';
 export { FakeAuth } from './mocks/auth';
 export { mockFirebase } from './mocks/firebase';
 export { mockGoogleCloudFirestore } from './mocks/googleCloudFirestore';
+export { mockReactNativeFirestore } from './mocks/reactNativeFirebaseFirestore';
